refactor: decompose the v0.10 fleet experience
