+++ docs/v0.10-multi-machine-monitoring.md
@@ -116,6 +116,10 @@ ssh -N -T -o BatchMode=yes -o ConnectTimeout=<bounded-seconds>
 ```
 
 Arguments are passed directly with no shell and no remote command.
+Enabled SSH registry entries must use distinct local ports. Before a new tunnel
+starts, the connector rejects a loopback port that is already accepting
+connections. After spawning `ssh`, it waits for that exact listener to become
+ready and keeps the whole operation inside the original request deadline.
 
 The host agent defaults to `127.0.0.1:4311`. It requires
 `GROK_UI_AGENT_TOKEN`; `GROK_UI_AGENT_HOST`, `GROK_UI_AGENT_PORT`, and
@@ -127,6 +131,11 @@ and file use `0700` and `0600` permissions. A malformed registry, invalid
 persisted host, or future schema version is preserved rather than overwritten;
 the central UI loads with an empty fleet plus `registryError`, and registry
 mutations remain blocked until the file is repaired.
+
+Registry mutations are transactional: they validate and persist a candidate
+map before exposing it in memory. Failed writes leave both memory and disk on
+the previous version, do not poison later writes, and preserve atomic
+temporary-file replacement.
 
 The polling interval and freshness thresholds are reported in every central
 fleet snapshot. Freshness is `fresh` for less than 10 seconds since success,
@@ -187,6 +196,28 @@ seconds, and `expired` at 45 seconds.
 - Run a 75-second multi-host soak, packed-install interoperability smoke,
   privacy and security checks, and `npm audit`.
 
+### Post-review architecture hardening
+
+Independent Grok and Claude reviews of the first candidate identified several
+cross-layer risks. The remediation keeps the same product scope while making
+the boundaries explicit:
+
+- the central server and host agent share session-row projection code, while
+  the agent composes current usage observations without becoming another
+  durable-state writer
+- refresh coalesces with active polling, unchanged snapshots do not force SSE
+  frames, and aggregate compaction uses per-host size deltas
+- registry transactions recover from failed writes and reject conflicting SSH
+  forwarding ports
+- the connector validates fixed paths before any fetch, waits for SSH
+  forwarding readiness, rejects redirects, and bounds streamed bodies
+- the browser validates and caps remote session detail before rendering it
+- Privacy Mode covers session-detail model labels and registry errors, while
+  status narratives retain safe generic operational meaning
+- `FleetView.tsx` owns page coordination only; selectors, status, the registry
+  editor, five telemetry panels, and Fleet CSS have separate owners under
+  `src/views/fleet/` and `src/styles/fleet.css`
+
 ## Requirement completion matrix
 
 `Complete` means both the implementation and current runtime evidence were
@@ -195,11 +226,11 @@ reviewed in this branch. A file reference by itself was not treated as proof.
 | Requirement | Implementation evidence | Automated/runtime evidence | Status |
 | --- | --- | --- | --- |
 | Lightweight host agent | `server/host-agent.ts`, `server/host-agent-entry.ts`, `bin/grok-ui.mjs` | Agent unit tests and packaged-agent authentication/method smoke passed | **Complete** |
-| Persistent central registry and fleet health view | `server/fleet-registry.ts`, `server/fleet-monitor.ts`, `server/index.ts`, `src/views/FleetView.tsx` | Registry/monitor tests, packaged restart, and Fleet browser journey passed | **Complete** |
+| Persistent central registry and fleet health view | `server/fleet-registry.ts`, `server/fleet-monitor.ts`, `server/index.ts`, `src/views/FleetView.tsx`, `src/views/fleet/` | Registry/monitor tests, packaged restart, and Fleet browser journey passed | **Complete** |
 | SSH- and Tailscale-oriented connectivity | `server/fleet-connectors.ts`, constrained registry parsing in `server/fleet-registry.ts` | Fixed-tunnel/destination tests and packaged loopback topology passed | **Complete** |
 | Identity, versions, capabilities, health, latency, last seen, freshness | `server/fleet-protocol.ts`, `server/fleet-monitor.ts`, `server/types.ts` | Protocol/monitor tests and Fleet browser assertions passed | **Complete** |
 | Read-only remote sessions, workflows, runtime state, and usage | Agent and central routes in `server/host-agent.ts` and `server/index.ts`; host namespacing in `server/fleet-protocol.ts` | Protocol/agent/monitor tests and remote browser inspection passed | **Complete** |
-| Explicit connecting, healthy, degraded, stale, offline, incompatible, unauthorized, and unavailable behavior | State machine in `server/fleet-monitor.ts`; narratives in `src/views/FleetView.tsx` | Deterministic unit cases plus multi-host and 45-second transition browser cases passed | **Complete** |
+| Explicit connecting, healthy, degraded, stale, offline, incompatible, unauthorized, and unavailable behavior | State machine in `server/fleet-monitor.ts`; narratives in `src/views/fleet/status.tsx` | Deterministic unit cases plus multi-host and 45-second transition browser cases passed | **Complete** |
 | No remote control, shell, arbitrary probing, or destructive surface | Fixed-path connector, GET/HEAD agent gate, stripped workflow controls, read-only Fleet panels | Negative agent/package cases and browser control-absence assertions passed | **Complete** |
 | Local-first behavior and reconnect/interruption reliability preserved | Existing local monitor, ACP, runtime, usage, and SSE paths retained in `server/index.ts` | Existing browser suite and exact 75-second ACP/fleet soak passed | **Complete** |
 | Privacy Mode and authentication preserved across fleet data | Dedicated agent bearer gate; tokens omitted by `publicHostConfig`; host and endpoint aliases in `src/privacy.tsx` | Agent/privacy/client tests plus browser privacy persistence passed | **Complete** |
@@ -209,20 +240,21 @@ reviewed in this branch. A file reference by itself was not treated as proof.
 
 ## Final verification ledger
 
-These results were recorded from the isolated v0.10 worktree after the global
-four-read concurrency regression was fixed and retained as a test case.
+These results were recorded from the isolated v0.10 worktree after the
+post-review correctness hardening and UI decomposition. The global four-read
+concurrency regression remains retained as a test case.
 
 | Gate | Command | Result |
 | --- | --- | --- |
 | Install | `npm ci --ignore-scripts` | **Passed** — 182 packages installed; 0 vulnerabilities |
 | Type safety | `npm run check` | **Passed** — client and server TypeScript projects |
-| Unit and integration | `npm test` | **Passed** — 57 tests in 16 files |
-| Production build and package audit | `npm run verify` | **Passed** — checks, tests, production build, 48-file/1.63 MiB package and privacy scan |
+| Unit and integration | `npm test` | **Passed** — 66 tests in 17 files |
+| Production build and package audit | `npm run verify` | **Passed** — checks, tests, production build, 49-file/1.64 MiB package and privacy scan |
 | Browser journey | `npm run test:e2e` | **Passed** — 19 Chromium scenarios, including 3 fleet scenarios |
 | Sustained monitoring | `npm run test:soak` | **Passed** — 75 seconds, 8 mock hosts, maximum 4 concurrent requests, stale/reconnect and ACP interruption confirmed |
 | Packed installation | `npm run test:package` | **Passed** — packaged agent/central interoperability, read-only gate, private persistence and restart |
 | Dependency audit | `npm audit --omit=dev --audit-level=high` | **Passed** — 0 vulnerabilities |
-| Working-tree audit | `git status --short` and review of the final diff | **Passed** — isolated branch contains only the intended v0.10 implementation and evidence updates |
+| Source and artifact audit | `git diff --check`, status, package manifest, and final diff review | **Passed** — only intended source, tests, and documentation; no generated or sensitive artifact included |
 
 ## Release gate
