+++ src/main.tsx
@@ -3,6 +3,7 @@ import { createRoot } from 'react-dom/client'
 import '@fontsource-variable/syne'
 import '@fontsource/ibm-plex-mono/400.css'
 import './styles.css'
+import './styles/fleet.css'
 import App from './App'
 import { ErrorBoundary } from './ErrorBoundary'
